Added interpolation of nans
